Unset PRIVACY_CONFIG before setting up a cluster
Cluster setup fails when the environment variable PRIVACY_CONFIG is set -- when
we tell geth to generate a new account in a temporary datadir, it tries to read
a constellation config which doesn't exist.

Currently to run geth with privacy enabled, we must set this environment
variable -- but this means that the variable is set for the remainder of the
lifetime of the process.

By clearing this environment variable before setting up the cluster, we can now
successfully set up a cluster after we have run a geth with privacy.